A couple of file utilities
* FileLock

This is a cooperative file lock that relies on the creation of a .lock
file to acquire the lock.

This is meant to be used to protect reading/writing of internal state
against concurrent executions of the tool. A similar mechanism is used
by `git` itself.

* AtomicFileWriter

Builds on top of FileLock, and implements the write-to-temp-then-rename
dance to make file updates atomic.

Once again, this is meant to be used to manage internal state and
protects concurrent executions of the tool.

Both of these should be supported cross major platforms, but require a
fairly recent version of `go`, and Windows 10.

* TestInDir

Also included in here, TestInDir is a test helper to run tests in a
temporary directory that is cleared at the end of the test. It also
includes support for redefining the HOME directory for the duration of
the test.