Defining the interface to be modelled
Stoic represents the engine. It will process configuration files,
maintain state, and generally orchestrate the setup, update and
execution of configured tools.

tool.Getter offers a mechanism to keep track and fetch upstream versions
of a project, and is also responsible for materializing releases on the
filesystem.

tool.Runner is responsible for taking a materialized release, performing
any necessary setup, and executing tools.

Also included, the configuration file format for defining tools.