Move tool.Tool to stoic.Tool
To avoid cyclic depedencies between tool and format, tool.Tool was moved
to the top-level as stoic.Tool, where it integrates closely with
stoic.Stoic.

The configuration format in format.ToolConfig was made dumber, and no
longer mangles the Enpoint URL. This is now handled by the engine, as
part of generating stoic.Tool instances.

Getter and Runner constructors were rewired on top to the new stoic.Tool
construct that grants access to tool configuration and some of its
state.

The engine's RunTool implementation was also updated to benefit from the
stoic.Tool abstraction.

In the process, the RunTool algorithm was simplified by dropping some
repetitive code and re-structuring in terms of 1) deciding on a version,
then 2) preparing a checkout, if needed, and 3) running the tool. With
this approach.

As part of deciding the version to run, a decision is made on whether to
fetch updates from the remote. If there isn't a current checkout, the
upstream will be fetched, even if it was previously decided not to
update. This allows for more frequent updates, if checking out the
upstream version has previously failed.

Otherwise, when a version is picked (and fetched), we make every effort
to service it locally by finding an existing checkout or creating a new
one specifically for that version.

One use case that is currently not well supported is the case where
release artifacts have been stored and evicted from the Cache, but there
isn't a checkout available for the tool. In the future, we should move
handling of the Cache to the engine, and out of the Getters, so the
engine can both produce checkouts straight from the cache, bypassing the
Getter, and re-fetch artifacts if they've been evicted.