Don't clobber a command's Stdout with setup output
This allows the output of a wrapped command to be piped/process without
undue unexpected content from setup commands.

For now, this is a policy that Getter and Runner must voluntarily adhere
to; in the future the engine might define specific streams for each use,
so this can be centrally managed and build logs kept around for future
reference.