Introducing the "go-get" Getter, and the "go-build" Runner
This allows running go tools from source. No attempt is made to
fetch/update dependencies, so they're expected to be maintained in the
repository, e.g., under a `vendor/` folder.

For `go get`, this pulls in "golang.org/x/tools/go/vcs" to find the VCS,
and repository root, but otherwise defers to the "git" Getter to track
the upstream. This also means that only dependencies hosted in "git" are
supported, for now.